rename to polynomial
