+++ threeML/utils/time_series/polynomial.py
